lab 2 finished

+++ lab2/src/main/java/com.lab2/repository/InMemoryRepositoryPrimitive.java
@@ -4,11 +4,11 @@ package com.lab2.repository;
  * Created by Cosmin on 10/30/2017.
  */
 public interface InMemoryRepositoryPrimitive {
-    void add(long e);
+    void add(int e);
 
-    boolean contains(long e);
+    boolean contains(int e);
 
-    void remove(long e);
+    void remove(int e);
 
     void clear();
 }
